Mms: Fixes mms automatic retrieval (1/2)
This patch uses the properties mms_auto_retrieval and mms_retrieval_during_roaming
which are already defined in the Mms app to set system-wide properties so that
automatic retrieve and retrieve on roaming are properly enforced and working.

The changes to the framework are in I824c7b48.

Patchset 1: Initial commit
Patchset 2: Fixed whitespace error
Patchset 3: Fixed comment..

Change-Id: Ibfcbcec841530dcee2db5cc7000e3131e2db1e5d