Mms: reload default drawable if it changed during the instance
Mms is holding the default contact drawable in a static drawable. This
has the side effect that when the drawable changes (eg because of an
configurationchange (in our special case TRDS) that the old drawable is
still holded and it does not update to the current actual one.

To avoid this make a fast compare with getConstantState and if it is
different refetch the default drawable and assign it to the static
drawable.

This fixes reported issue when Mms is open and the user switches it
and default images does not switch.

Change-Id: Id369588d675643e8ea8d443a39849ba9eeef0560