Mms: Fix the draft Mms always show as SIM2.
When insert draft messages into database, the "phone_id" value is not
appointed and "-1" is set as the dafault value.
When displaying the draft in folder mode, the "phone_id" value(-1) can
not match the SIM 1 case and run into SIM2 case by default.

Do not show sim indicator when displaying draft messages.

CRs-Fixed: 738001
Change-Id: If1f838f832b5c24923a5c52fe97786d691760a91